requirements.txt
# --------------------------------------------
# Dependencias principales de tu proyecto Django
# --------------------------------------------
aiofiles==0.8.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
apache-atlas==0.0.16
asgiref==3.7.2
asttokens==3.0.0
attrs==24.2.0
Authlib==1.3.0
blinker==1.9.0
boto3==1.35.82
botocore==1.35.82
cached-property==1.5.2
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
coverage==7.4.4
cryptography==42.0.5
dash==3.0.1
decorator==5.2.1
dependency-injector==4.41.0
deprecation==2.1.0
Django==5.0.2
django-admin-interface==0.28.8
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
django-keycloak-auth==0.9.5
django-nested-admin==4.1.1
django-redis==5.4.0
django-reversion==5.1.0
django-sortable==0.3.1
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
dnspython==2.6.1
docx==0.2.4
drf-extensions==0.7.1
drf-spectacular==0.27.1
drf-yasg==1.21.10
ecdsa==0.18.0
elastic-transport==8.15.0
elasticsearch==8.15.1
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.110.0
fastjsonschema==2.21.1
Flask==3.0.3
freetype-py==2.5.1
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.67.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0

# -------------------------------------------------
# Aquí corrigimos la versión de httpx para evitar conflicto con keycloak
# keycloak 3.1.4 requiere httpx>=0.26.0,<0.27.0
# -------------------------------------------------
httpx==0.26.0

httptools==0.6.1
hyperframe==6.0.1
idna==3.6
imageio==2.37.0
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.0.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
jwcrypto==1.5.6

# Versión de keycloak que tú estabas usando (3.1.4)
keycloak==3.1.4

lazy_loader==0.4
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mssql-django==1.5
multidict==6.1.0
narwhals==1.32.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
oauthlib==3.1.1
open3d==0.19.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
prompt_toolkit==3.0.50
protobuf==3.20.3
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.5.1
pycparser==2.21

# Estas tres líneas las comentamos porque no son necesarias o fallan en Linux
# python-slugify==8.0.4
# pytz==2024.1
# pywin32==310

pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==3.9.1
python-monkey-business==1.1.0
python-multipart==0.0.17
PyYAML==6.0.1
pyzeebe==3.0.4
redis==5.0.8
referencing==0.33.0
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3986==2.0.0
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.4
scikit-image==0.25.2
scipy==1.15.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.36.3
text-unidecode==1.3
tifffile==2025.6.1
traitlets==5.14.3
trimesh==4.6.5
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.4.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.28.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.6
whitenoise==6.6.0
widgetsnbextension==4.0.13
XlsxWriter==3.2.0
yarl==1.13.1
zeebe-grpc==1.3.7
zipp==3.21.0
